[PR/MERGE] PWT, lecture, Product 테이블 계층구조 (부모/자식 테이블 계층구조)

+++ src/main/java/com/devloop/common/apipayload/dto/ProjectWithTutorResponseDto.java
@@ -8,6 +8,7 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Builder
@@ -16,7 +17,7 @@ import java.time.LocalDateTime;
 public class ProjectWithTutorResponseDto {
     private final Long id;
     private final String title;
-    private final Integer price;
+    private final BigDecimal price;
     private final ProjectWithTutorStatus status;
     private final LocalDateTime deadline;
     private final Integer maxParticipants;
